# openapi.yaml
openapi: 3.0.3
info:
  title: CubeOS API
  description: |
    CubeOS unified API for managing Docker services, network configuration,
    VPN connections, and storage mounts on Raspberry Pi.
  version: 2.0.0
  contact:
    name: CubeOS
    url: https://cubeos.app
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://api.cubeos.cube/api/v1
    description: Internal CubeOS network
  - url: http://10.42.24.1:6010/api/v1
    description: Direct IP access
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  example: cubeos
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  expires_at:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================================
  # Apps (Unified)
  # ============================================================
  /apps:
    get:
      tags: [Apps]
      summary: List all apps
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [system, platform, network, ai, user]
        - name: status
          in: query
          schema:
            type: string
            enum: [running, stopped]
        - name: enabled
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of apps
          content:
            application/json:
              schema:
                type: object
                properties:
                  apps:
                    type: array
                    items:
                      $ref: '#/components/schemas/App'
    post:
      tags: [Apps]
      summary: Install app
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstallAppRequest'
      responses:
        '201':
          description: App installed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/App'
        '409':
          $ref: '#/components/responses/Conflict'
  /apps/{name}:
    get:
      tags: [Apps]
      summary: Get app details
      parameters:
        - $ref: '#/components/parameters/AppName'
      responses:
        '200':
          description: App details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/App'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Apps]
      summary: Uninstall app
      parameters:
        - $ref: '#/components/parameters/AppName'
        - name: keep_data
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: App uninstalled
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
  /apps/{name}/start:
    post:
      tags: [Apps]
      summary: Start app
      parameters:
        - $ref: '#/components/parameters/AppName'
      responses:
        '200':
          description: App started
        '404':
          $ref: '#/components/responses/NotFound'
  /apps/{name}/stop:
    post:
      tags: [Apps]
      summary: Stop app
      parameters:
        - $ref: '#/components/parameters/AppName'
      responses:
        '200':
          description: App stopped
        '404':
          $ref: '#/components/responses/NotFound'
  /apps/{name}/restart:
    post:
      tags: [Apps]
      summary: Restart app
      parameters:
        - $ref: '#/components/parameters/AppName'
      responses:
        '200':
          description: App restarted
        '404':
          $ref: '#/components/responses/NotFound'
  /apps/{name}/enable:
    post:
      tags: [Apps]
      summary: Enable app (start on boot)
      parameters:
        - $ref: '#/components/parameters/AppName'
      responses:
        '200':
          description: App enabled
  /apps/{name}/disable:
    post:
      tags: [Apps]
      summary: Disable app (don't start on boot)
      parameters:
        - $ref: '#/components/parameters/AppName'
      responses:
        '200':
          description: App disabled
  /apps/{name}/logs:
    get:
      tags: [Apps]
      summary: Get app logs
      parameters:
        - $ref: '#/components/parameters/AppName'
        - name: lines
          in: query
          schema:
            type: integer
            default: 100
        - name: since
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: App logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      type: string
...
  /profiles:
    get:
      tags: [Profiles]
      summary: List profiles
      responses:
        '200':
          description: List of profiles
          content:
            application/json:
              schema:
                type: object
                properties:
                  profiles:
                    type: array
                    items:
                      $ref: '#/components/schemas/Profile'
                  active_profile:
                    type: string
    post:
      tags: [Profiles]
      summary: Create custom profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProfileRequest'
      responses:
        '201':
          description: Profile created
  /profiles/{name}:
    get:
      tags: [Profiles]
      summary: Get profile details
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Profile details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
...
  /network/mode:
    post:
      tags: [Network]
      summary: Set network mode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetNetworkModeRequest'
      responses:
        '200':
          description: Network mode changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NetworkStatus'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /network/wifi/connect:
    post:
      tags: [Network]
      summary: Connect to upstream WiFi
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ssid, password]
              properties:
                ssid:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Connected to WiFi
        '500':
          $ref: '#/components/responses/InternalError'

  # ============================================================
  # VPN
  # ============================================================
...
  /mounts:
    get:
      tags: [Mounts]
      summary: List all mounts
      responses:
        '200':
          description: List of mounts
          content:
            application/json:
              schema:
                type: object
                properties:
                  mounts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Mount'
    post:
      tags: [Mounts]
      summary: Add mount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMountRequest'
      responses:
        '201':
          description: Mount added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mount'
...
  /mounts/{name}:
    get:
      tags: [Mounts]
      summary: Get mount details
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Mount details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mount'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Mounts]
      summary: Delete mount
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Mount deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    AppName:
      name: name
      in: path
      required: true
      schema:
        type: string
      description: App name (lowercase, no spaces)
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden - system app protected
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Conflict - resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
    App:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        display_name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [system, platform, network, ai, user]
        category:
          type: string
        enabled:
          type: boolean
        tor_enabled:
          type: boolean
        vpn_enabled:
          type: boolean
        ports:
          type: array
          items:
            $ref: '#/components/schemas/Port'
        fqdns:
          type: array
          items:
            $ref: '#/components/schemas/FQDN'
        status:
          $ref: '#/components/schemas/AppStatus'
    Port:
      type: object
      properties:
        port:
          type: integer
        protocol:
          type: string
          enum: [tcp, udp]
        description:
          type: string
        is_primary:
          type: boolean
    FQDN:
      type: object
      properties:
        fqdn:
          type: string
        subdomain:
          type: string
    AppStatus:
      type: object
      properties:
        running:
          type: boolean
        health:
          type: string
          enum: [healthy, unhealthy, starting, unknown]
        replicas:
          type: string
          example: "1/1"
    InstallAppRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        display_name:
          type: string
        source:
          type: string
          enum: [cubeos, casaos, custom]
        store_app_id:
          type: string
    Profile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        display_name:
          type: string
        description:
          type: string
        is_active:
          type: boolean
        is_system:
          type: boolean
        apps:
          type: array
          items:
            type: object
            properties:
              app_id:
                type: integer
              app_name:
                type: string
              enabled:
                type: boolean
    CreateProfileRequest:
      type: object
      required: [name, display_name]
      properties:
        name:
          type: string
        display_name:
          type: string
        apps:
          type: array
          items:
            type: object
            properties:
              app_name:
                type: string
              enabled:
                type: boolean
    NetworkStatus:
      type: object
      properties:
        mode:
          type: string
          enum: [offline, online_eth, online_wifi]
        internet:
          type: boolean
        ap:
          type: object
          properties:
            ssid:
              type: string
            interface:
              type: string
            clients:
              type: integer
        upstream:
          type: object
          properties:
            interface:
              type: string
            ip:
              type: string
            gateway:
              type: string
        subnet:
          type: string
        gateway_ip:
          type: string
    SetNetworkModeRequest:
      type: object
      required: [mode]
      properties:
        mode:
          type: string
          enum: [offline, online_eth, online_wifi]
        ssid:
          type: string
          description: Required for online_wifi mode
        password:
          type: string
          description: Required for online_wifi mode
...
    Mount:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [smb, nfs]
        remote_path:
          type: string
        local_path:
          type: string
        username:
          type: string
        is_mounted:
          type: boolean
        auto_mount:
          type: boolean
        created_at:
          type: string
          format: date-time
    AddMountRequest:
      type: object
      required: [name, type, remote_path]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [smb, nfs]
        remote_path:
          type: string
          description: "SMB: //server/share, NFS: server:/path"
        username:
          type: string
        password:
          type: string
        options:
          type: string
        auto_mount:
          type: boolean
...